Added big_welcome_banner.png; moved images to images folder; corrected imports accordingly
